FIXME ROUGH day 23 part 02
Mostly copy/paste of part 01 with minor changes to cup label integer
sizes (u8 -> u32). This actually works and only takes a second or so to
solve (thanks rust) but there is definitely a more efficient solution.
Looking at it, I'm not sure that the elaborate HashMap/LinkedList data
structure is necessary and is *certainly* hurting performance. Is there
another way to model links between labels without *physically*
representing the links as pointers to nodes?

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>